fix: update HeroSection and refactor code
