+++ assets/lang/en/password-generator.json
@@ -3,6 +3,8 @@
     "title": "Password Generator",
     "subtitle": "Generate highly secure and unique passwords in an instant!",
     "info": "Internxt does not store or transmit passwords.",
+    "copy": "Copy password",
+    "generate": "Generate new",
     "password": {
       "title": "Password",
       "length": "Password length",
@@ -21,13 +23,7 @@
         "numbers": "Numbers",
         "separator": {
           "title": "Separate words by ",
-          "options": {
-            "comma": ",",
-            "space": "Space",
-            "dot": ".",
-            "interrogation": "?",
-            "dollars": "$"
-          }
+          "options": [",", "\n", ".", "?", "$", "-"]
         }
       }
     }
@@ -41,7 +37,7 @@
       ]
     },
     "secondSection": {
-      "title": "Say goodbye to weak passwords with Internx’s secure password generator",
+      "title": "Say goodbye to weak passwords with Internxt’s secure password generator",
       "description": [
         "While it is possible to create strong passwords yourself, it can be a time-consuming and complicated process. With Internxt, a strong password or passkey is just one click away.",
         "We developed our unique password generator that uses advanced algorithms to create unique passwords which follow these rules.",
